requirements.txt
openai
python-dotenv
chromadb==0.4.24
sentence-transformers
torch
numpy<2.0
redis
